# slight increase in performance