--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
     args: ["--profile", "black"]
     name: isort (python)
 - repo: https://github.com/psf/black
-  rev: 21.8b0
+  rev: 22.3.0
   hooks:
   - id: black
     args: ["--target-version", "py36"]