@deedy5 2022-05-26
lint error:
```python
ImportError: cannot import name '_unicodefun' from 'click'
```
https://github.com/psf/black/issues/2964#issuecomment-1080974737